Jewels and Stones
This Java solution addresses the "Jewels and Stones" problem on LeetCode. The goal of the problem is to determine the number of stones that are also jewels given a string representing the types of jewels (J) and a string representing the stones (S). The solution utilizes a straightforward approach, iterating through each character in the stones string and checking if it is contained within the jewels string. If a stone is also a jewel, it increments the count. The runtime performance of this solution is relatively efficient, with a runtime beating 62.34% of Java submissions and memory usage outperforming 93.93% of submissions.